requirements.txt
google-genai==0.4.1
python-dotenv==1.0.0
requests==2.32.0
PyYAML==6.0.1